+++ Source/Harmony/ReputationWidget.cs
@@ -28,41 +28,6 @@ namespace RP0
                 return false;
             }
 
-            internal static void CreateConfidenceWidget(GameObject confidenceWidgetObj)
-            {
-                confidenceWidgetObj.name = "ConfidenceWidget";
-
-                GameObject.Destroy(confidenceWidgetObj.GetComponent<ReputationWidget>());
-
-                var frameImage = (Image)confidenceWidgetObj.GetComponentInChildren(typeof(Image));
-                frameImage.sprite = Sprite.Create(GameDatabase.Instance.GetTexture("RP-0/Resources/confidence_background", false), frameImage.sprite.rect, frameImage.sprite.pivot);
-
-                var img = GameObject.Instantiate(new GameObject("Background"), confidenceWidgetObj.transform, worldPositionStays: false).AddComponent<Image>();
-                img.color = new Color32(58, 58, 63, 255);
-                img.rectTransform.anchorMin = frameImage.rectTransform.anchorMin;
-                img.rectTransform.anchorMax = frameImage.rectTransform.anchorMax;
-                img.rectTransform.anchoredPosition = frameImage.rectTransform.anchoredPosition;
-                img.rectTransform.sizeDelta = ((RectTransform)confidenceWidgetObj.transform).sizeDelta;    // No idea why the frame image transform is larger than the component itself
-
-                var textComp = GameObject.Instantiate(new GameObject("Text"), confidenceWidgetObj.transform, worldPositionStays: false).AddComponent<TextMeshProUGUI>();
-                textComp.alignment = TextAlignmentOptions.Right;
-                textComp.color = XKCDColors.KSPBadassGreen;
-                textComp.fontSize = 22;
-                textComp.rectTransform.localPosition = new Vector3(-9, -1, 0);
-                textComp.fontStyle = FontStyles.Bold;
-
-                var confidenceWidget = confidenceWidgetObj.AddComponent<ConfidenceWidget>();
-                confidenceWidget.text = textComp;
-                confidenceWidget.DelayedStart();
-
-                // Add tooltip
-                var tooltip = confidenceWidgetObj.AddComponent<TooltipController_TextFunc>();
-                var prefab = AssetBase.GetPrefab<Tooltip_Text>("Tooltip_Text");
-                tooltip.prefab = prefab;
-                tooltip.getStringAction = GetTooltipTextConf;
-                tooltip.continuousUpdate = true;
-            }
-
             [HarmonyPrefix]
             [HarmonyPatch("DelayedStart")]
             internal static bool Prefix_DelayedStart(ReputationWidget __instance)
@@ -72,7 +37,7 @@ namespace RP0
                 GameObject.DestroyImmediate(__instance.gameObject.transform.Find("circularGauge").gameObject);
 
                 // Create the Confidence widget
-                CreateConfidenceWidget(GameObject.Instantiate(__instance.gameObject, __instance.transform.parent, worldPositionStays: false));
+                ConfidenceWidget.CreateConfidenceWidget(GameObject.Instantiate(__instance.gameObject, __instance.transform.parent, worldPositionStays: false));
 
                 var frameImage = (Image)__instance.gameObject.GetComponentInChildren(typeof(Image));
                 frameImage.sprite = Sprite.Create(GameDatabase.Instance.GetTexture("RP-0/Resources/rep_background", false), frameImage.sprite.rect, frameImage.sprite.pivot);
@@ -99,11 +64,6 @@ namespace RP0
                 tooltip.continuousUpdate = false;
 
                 return true;
-            }
-
-            private static string GetTooltipTextConf()
-            {
-                return Localizer.Format("#rp0ConfidenceWidgetTooltip", Confidence.AllConfidenceEarned.ToString("N0"));
             }
 
             private static string GetTooltipTextRep()
